+++ src/dune_engine/dune_project.ml
@@ -160,8 +160,8 @@ type t =
   ; implicit_transitive_deps : bool
   ; wrapped_executables : bool
   ; dune_version : Dune_lang.Syntax.Version.t
-  ; use_standard_c_and_cxx_flags : bool
   ; generate_opam_files : bool
+  ; use_standard_c_and_cxx_flags : bool option
   ; file_key : File_key.t
   ; dialects : Dialect.DB.t
   ; explicit_js_mode : bool
@@ -190,9 +190,9 @@ let file_key t = t.file_key
 
 let implicit_transitive_deps t = t.implicit_transitive_deps
 
-let use_standard_c_and_cxx_flags t = t.use_standard_c_and_cxx_flags
-
 let generate_opam_files t = t.generate_opam_files
+
+let use_standard_c_and_cxx_flags t = t.use_standard_c_and_cxx_flags
 
 let dialects t = t.dialects
 
@@ -211,8 +211,8 @@ let to_dyn
     ; implicit_transitive_deps
     ; wrapped_executables
     ; dune_version
-    ; use_standard_c_and_cxx_flags
     ; generate_opam_files
+    ; use_standard_c_and_cxx_flags
     ; file_key
     ; dialects
     ; explicit_js_mode
@@ -233,8 +233,8 @@ let to_dyn
     ; ("implicit_transitive_deps", bool implicit_transitive_deps)
     ; ("wrapped_executables", bool wrapped_executables)
     ; ("dune_version", Dune_lang.Syntax.Version.to_dyn dune_version)
-    ; ("use_standard_c_and_cxx_flags", bool use_standard_c_and_cxx_flags)
     ; ("generate_opam_files", bool generate_opam_files)
+    ; ("use_standard_c_and_cxx_flags", option bool use_standard_c_and_cxx_flags)
     ; ("file_key", string file_key)
     ; ("dialects", Dialect.DB.to_dyn dialects)
     ; ("explicit_js_mode", bool explicit_js_mode)
@@ -609,8 +609,12 @@ let infer ~dir packages =
   ; extension_args
   ; parsing_context
   ; dune_version = lang.version
-  ; use_standard_c_and_cxx_flags = false
   ; generate_opam_files = false
+  ; use_standard_c_and_cxx_flags =
+      ( if lang.version < (3, 0) then
+        None
+      else
+        Some true )
   ; file_key
   ; dialects = Dialect.DB.builtin
   ; explicit_js_mode
@@ -691,13 +695,13 @@ let parse ~dir ~lang ~opam_packages ~file ~dir_status =
              "It is useless since the Merlin configurations are not ambiguous \
               anymore."
            loc lang.syntax (2, 8) ~what:"This field"
-     and+ use_standard_c_and_cxx_flags =
-       field_o_b "use_standard_c_and_cxx_flags"
-         ~check:(Dune_lang.Syntax.since Stanza.syntax (2, 8))
      and+ () = Dune_lang.Versioned_file.no_more_lang
      and+ generate_opam_files =
        field_o_b "generate_opam_files"
          ~check:(Dune_lang.Syntax.since Stanza.syntax (1, 10))
+     and+ use_standard_c_and_cxx_flags =
+       field_o_b "use_standard_c_and_cxx_flags"
+         ~check:(Dune_lang.Syntax.since Stanza.syntax (2, 8))
      and+ dialects =
        multi_field "dialect"
          ( Dune_lang.Syntax.since Stanza.syntax (1, 11)
@@ -806,14 +810,17 @@ let parse ~dir ~lang ~opam_packages ~file ~dir_status =
          ~default:(strict_package_deps_default ~lang)
      in
      let dune_version = lang.version in
-     let use_standard_c_and_cxx_flags =
-       Option.value ~default:false use_standard_c_and_cxx_flags
-     in
      let explicit_js_mode =
        Option.value explicit_js_mode ~default:(explicit_js_mode_default ~lang)
      in
      let generate_opam_files =
        Option.value ~default:false generate_opam_files
+     in
+     let use_standard_c_and_cxx_flags =
+       match use_standard_c_and_cxx_flags with
+       | None when dune_version >= (3, 0) -> Some false
+       | None -> None
+       | some -> some
      in
      let cram =
        match cram with
@@ -858,8 +865,8 @@ let parse ~dir ~lang ~opam_packages ~file ~dir_status =
      ; implicit_transitive_deps
      ; wrapped_executables
      ; dune_version
-     ; use_standard_c_and_cxx_flags
      ; generate_opam_files
+     ; use_standard_c_and_cxx_flags
      ; dialects
      ; explicit_js_mode
      ; format_config