+++ src/dune_rules/foreign_rules.ml
@@ -75,19 +75,53 @@ let include_dir_flags ~expander ~dir (stubs : Foreign.Stubs.t) =
 
 let build_c ~kind ~sctx ~dir ~expander ~include_flags (loc, src, dst) =
   let ctx = Super_context.context sctx in
+  let project = Super_context.find_scope_by_dir sctx dir |> Scope.project in
+  let use_standard_flags = Dune_project.use_standard_c_and_cxx_flags project in
   let base_flags =
     let cfg = ctx.ocaml_config in
     match kind with
-    | Foreign_language.C ->
-      List.concat
-        [ Ocaml_config.ocamlc_cflags cfg
-        ; Ocaml_config.ocamlc_cppflags cfg
-        ; Fdo.c_flags ctx
-        ]
+    | Foreign_language.C -> (
+      match use_standard_flags with
+      | None
+      | Some false ->
+        (* In dune < 2.8 flags from ocamlc_config are always added *)
+        List.concat
+          [ Ocaml_config.ocamlc_cflags cfg
+          ; Ocaml_config.ocamlc_cppflags cfg
+          ; Fdo.c_flags ctx
+          ]
+      | Some true -> Fdo.c_flags ctx )
     | Foreign_language.Cxx -> Fdo.cxx_flags ctx
   in
   let with_user_and_std_flags =
     let flags = Foreign.Source.flags src in
+    (* DUNE3 will have [use_standard_c_and_cxx_flags] enabled by default. To
+       guide users toward this change we emit a warning when dune_lang is >=
+       1.8, [use_standard_c_and_cxx_flags] is not specified in the
+       [dune-project] file (thus defaulting to [true]), the [:standard] set of
+       flags has been overriden and we are not in a vendored project *)
+    let has_standard = Ordered_set_lang.Unexpanded.has_standard flags in
+    let is_vendored =
+      match Path.Build.drop_build_context dir with
+      | Some src_dir -> Dune_engine.File_tree.is_vendored src_dir
+      | None -> false
+    in
+    if
+      Dune_project.dune_version project >= (2, 8)
+      && Option.is_none use_standard_flags
+      && (not is_vendored) && not has_standard
+    then
+      User_warning.emit ~loc
+        [ Pp.text
+            "The flag set for these foreign sources overrides the `:standard` \
+             set of flags. However the flags in this standard set are still \
+             added to the compiler arguments by Dune. This might cause \
+             unexpected issues. You can disable this warning by defining the \
+             option `(use_standard_c_and_cxx_flags <bool>)` in your \
+             `dune-project` file. Setting this option to `true` will \
+             effectively prevent Dune from silently adding c-flags to the \
+             compiler arguments which is the new recommended behaviour."
+        ];
     Super_context.foreign_flags sctx ~dir ~expander ~flags ~language:kind
     |> Build.map ~f:(List.append base_flags)
   in